+++ R/assertthat.R
@@ -5,7 +5,7 @@
 #' @param x object to test
 #' @param df supposed data frame to test
 #' @param variables supposed variables list to test
-#' @param include_boundaries (`logical`)\cr whether to include boundaries in `is_proportion`
+#' @param include_boundaries (`logical`)\cr whether to include boundaries when testing for proportions.
 #' @param ... a collection of objects to test.
 #' @return `flag` whether the assertion holds (`TRUE` or `FALSE`). When used inside
 #'   [assertthat::assert_that()] produces a meaningful error message.
@@ -122,17 +122,33 @@ on_failure(is_equal_length) <- function(call, env) {
 #' is_proportion(x = 0.3)
 #' is_proportion(x = 1.3)
 #' is_proportion(x = 0, include_boundaries = TRUE)
+#'
 is_proportion <- function(x, include_boundaries = FALSE) {
-  if (include_boundaries) {
-    is_numeric_single(x) &&
-      x >= 0 && x <= 1
-  } else {
-    is_numeric_single(x) &&
-      x > 0 && x < 1
-  }
+  is_numeric_single(x) &&
+    is_proportion_vector(x, include_boundaries = include_boundaries)
 }
 on_failure(is_proportion) <- function(call, env) {
   paste(deparse(call$x), "is not a proportion: number between 0 and 1")
+}
+
+#' @describeIn assertions Check whether `x` is a vector of proportions (numbers between 0 and 1).
+#' @export
+#' @examples
+#'
+#' # Check whether `x` is a vector of numbers between 0 and 1.
+#' is_proportion_vector(c(0.3, 0.1))
+#' is_proportion_vector(c(1.3, 0.1))
+#' is_proportion_vector(0, include_boundaries = TRUE)
+#'
+is_proportion_vector <- function(x, include_boundaries = FALSE) {
+  if (include_boundaries) {
+    all(x >= 0 & x <= 1)
+  } else {
+    all(x > 0 & x < 1)
+  }
+}
+on_failure(is_proportion_vector) <- function(call, env) {
+  paste(deparse(call$x), "is not a vector of proportions (numbers between 0 and 1)")
 }
 
 #' @describeIn assertions Check whether `x` is a valid factor (has levels and no empty string levels).
@@ -201,4 +217,20 @@ on_failure(all_elements_in_ref) <- function(call, env) {
   not_in_ref <- x[!(x %in% ref)]
 
   paste0("Some elements ", deparse(not_in_ref), " are not in the reference.")
+}
+
+#' @describeIn assertions Check whether rtables object `x` has the specified column names.
+#' @param col_names (`character`)\cr column names which should be present in the table.
+#'
+has_tabletree_colnames <- function(x, col_names) {
+  is(x, "VTableNodeInfo") &&
+    !is.null(names(x)) &&
+    all(col_names %in% names(x))
+}
+on_failure(has_tabletree_colnames) <- function(call, env) {
+  x <- eval(call$x, envir = env)
+  col_names <- eval(call$col_names, envir = env)
+  missing_col_names <- x[!(x %in% col_names)]
+
+  paste0("required column names ", deparse(missing_col_names), " are not found in table")
 }